+++ benchmarks/bench_algos.py
@@ -233,3 +233,9 @@ def bench_graph_degrees(gpubenchmark, anyGraphWithAdjListComputed):
 def bench_betweenness_centrality(gpubenchmark, anyGraphWithAdjListComputed):
     gpubenchmark(cugraph.betweenness_centrality,
                  anyGraphWithAdjListComputed, k=10, seed=123)
+
+
+def bench_edge_betweenness_centrality(gpubenchmark,
+                                      anyGraphWithAdjListComputed):
+    gpubenchmark(cugraph.edge_betweenness_centrality,
+                 anyGraphWithAdjListComputed, k=10, seed=123)