+++ python/cugraph/centrality/betweenness_centrality_wrapper.pyx
@@ -17,16 +17,24 @@
 # cython: language_level = 3
 
 from cugraph.centrality.betweenness_centrality cimport betweenness_centrality as c_betweenness_centrality
-from cugraph.structure.graph_new cimport *
-from cugraph.utilities.unrenumber import unrenumber
-from libcpp cimport bool
-from libc.stdint cimport uintptr_t
 from cugraph.structure import graph_new_wrapper
 from cugraph.structure.graph import DiGraph
+from cugraph.structure.graph_new cimport *
+from cugraph.utilities.unrenumber import unrenumber
+from libc.stdint cimport uintptr_t
+from libcpp cimport bool
 import cudf
-import rmm
 import numpy as np
 import numpy.ctypeslib as ctypeslib
+
+
+def get_output_df(input_graph, result_dtype):
+    number_of_vertices = input_graph.number_of_vertices()
+    df = cudf.DataFrame()
+    df['vertex'] = cudf.Series(np.zeros(number_of_vertices, dtype=np.int32))
+    df['betweenness_centrality'] = cudf.Series(np.zeros(number_of_vertices,
+                                                        dtype=result_dtype))
+    return df
 
 
 def betweenness_centrality(input_graph, normalized, endpoints, weight, k,
@@ -34,36 +42,29 @@ def betweenness_centrality(input_graph, normalized, endpoints, weight, k,
     """
     Call betweenness centrality
     """
-    cdef GraphCSRView[int, int, float] graph_float
-    cdef GraphCSRView[int, int, double] graph_double
+    cdef GraphCSRViewFloat graph_float
+    cdef GraphCSRViewDouble graph_double
+    cdef uintptr_t c_identifier = <uintptr_t> NULL
+    cdef uintptr_t c_betweenness = <uintptr_t> NULL
+    cdef uintptr_t c_vertices = <uintptr_t> NULL
+    cdef uintptr_t c_weight = <uintptr_t> NULL
 
     if not input_graph.adjlist:
         input_graph.view_adj_list()
 
-    [offsets, indices] = graph_new_wrapper.datatype_cast([input_graph.adjlist.offsets, input_graph.adjlist.indices], [np.int32])
+    df = get_output_df(input_graph, result_dtype)
 
-    num_verts = input_graph.number_of_vertices()
-    num_edges = input_graph.number_of_edges(directed_edges=True)
-
-    df = cudf.DataFrame()
-    df['vertex'] = cudf.Series(np.zeros(num_verts, dtype=np.int32))
-    df['betweenness_centrality'] = cudf.Series(np.zeros(num_verts, dtype=result_dtype))
-
-    cdef uintptr_t c_identifier = df['vertex'].__cuda_array_interface__['data'][0]
-    cdef uintptr_t c_betweenness = df['betweenness_centrality'].__cuda_array_interface__['data'][0]
-
-    cdef uintptr_t c_offsets = offsets.__cuda_array_interface__['data'][0]
-    cdef uintptr_t c_indices = indices.__cuda_array_interface__['data'][0]
-    cdef uintptr_t c_weight  = <uintptr_t> NULL
-    cdef uintptr_t c_vertices = <uintptr_t> NULL
+    c_identifier = df['vertex'].__cuda_array_interface__['data'][0]
+    c_betweenness = df['betweenness_centrality'].__cuda_array_interface__['data'][0]
 
     if weight is not None:
         c_weight = weight.__cuda_array_interface__['data'][0]
 
-    #FIXME: We could sample directly from a cudf array in the futur: i.e
-    #       c_vertices = vertices.__cuda_array_interface__['data'][0]
     if vertices is not None:
-        c_vertices =  np.array(vertices, dtype=np.int32).__array_interface__['data'][0]
+        # NOTE: Do not merge lines, c_vertices may end up pointing at the
+        #       wrong place the length of vertices increase.
+        np_verts =  np.array(vertices, dtype=np.int32)
+        c_vertices = np_verts.__array_interface__['data'][0]
 
     c_k = 0
     if k is not None:
@@ -74,8 +75,7 @@ def betweenness_centrality(input_graph, normalized, endpoints, weight, k,
     #       The current BFS requires the GraphCSR to be declared
     #       as <int, int, float> or <int, int double> even if weights is null
     if result_dtype == np.float32:
-        graph_float = GraphCSRView[int, int, float](<int*> c_offsets, <int*> c_indices,
-                                                <float*> NULL, num_verts, num_edges)
+        graph_float = get_graph_view[GraphCSRViewFloat](input_graph, False)
         # FIXME: There might be a way to avoid manually setting the Graph property
         graph_float.prop.directed = type(input_graph) is DiGraph
 
@@ -84,10 +84,9 @@ def betweenness_centrality(input_graph, normalized, endpoints, weight, k,
                                                          normalized, endpoints,
                                                          <float*> c_weight, c_k,
                                                          <int*> c_vertices)
-        graph_float.get_vertex_identifiers(<int*>c_identifier)
+        graph_float.get_vertex_identifiers(<int*> c_identifier)
     elif result_dtype == np.float64:
-        graph_double = GraphCSRView[int, int, double](<int*>c_offsets, <int*>c_indices,
-                                                  <double*> NULL, num_verts, num_edges)
+        graph_double = get_graph_view[GraphCSRViewDouble](input_graph, False)
         # FIXME: There might be a way to avoid manually setting the Graph property
         graph_double.prop.directed = type(input_graph) is DiGraph
 
@@ -96,12 +95,12 @@ def betweenness_centrality(input_graph, normalized, endpoints, weight, k,
                                                            normalized, endpoints,
                                                            <double*> c_weight, c_k,
                                                            <int*> c_vertices)
-        graph_double.get_vertex_identifiers(<int*>c_identifier)
+        graph_double.get_vertex_identifiers(<int*> c_identifier)
     else:
         raise TypeError("result type for betweenness centrality can only be "
                         "float or double")
 
-    #FIXME: For large graph renumbering produces a dataframe organized
+    # For large graph unrenumbering produces a dataframe organized
     #       in buckets, i.e, if they are 3 buckets
     # 0
     # 8191