+++ python/cugraph/tests/test_bfs.py
@@ -48,26 +48,6 @@ def prepare_test():
     gc.collect()
 
 
-# TODO: This is also present in test_betweenness_centrality.py
-#       And it could probably be used in SSSP also
-def build_graphs(graph_file, directed=True):
-    # cugraph
-    cu_M = utils.read_csv_file(graph_file)
-    # Get unsymmetrized/directed edges
-    cu_M = cu_M[cu_M['0'] <= cu_M['1']].reset_index(drop=True)
-    G = cugraph.DiGraph() if directed else cugraph.Graph()
-    G.from_cudf_edgelist(cu_M, source='0', destination='1')
-    G.view_adj_list()  # Enforce CSR generation before computation
-
-    # networkx
-    M = utils.read_csv_for_nx(graph_file)
-    M = M[M['0'] <= M['1']]
-    Gnx = nx.from_pandas_edgelist(M, create_using=(nx.DiGraph() if directed
-                                                   else nx.Graph()),
-                                  source='0', target='1')
-    return G, Gnx
-
-
 # =============================================================================
 # Functions for comparison
 # =============================================================================
@@ -98,7 +78,7 @@ def compare_bfs(graph_file, directed=True, return_sp_counter=False,
     Returns
     -------
     """
-    G, Gnx = build_graphs(graph_file, directed)
+    G, Gnx = utils.build_cu_and_nx_graphs(graph_file, directed)
     # Seed for reproducibility
     if isinstance(seed, int):
         random.seed(seed)
@@ -116,18 +96,18 @@ def compare_bfs(graph_file, directed=True, return_sp_counter=False,
     elif isinstance(seed, list):  # For other Verifications
         for start_vertex in seed:
             compare_func = _compare_bfs_spc if return_sp_counter else \
-                           _compare_bfs
+                _compare_bfs
             compare_func(G, Gnx, start_vertex)
     elif seed is None:  # Same here, it is only to run full checks
         for start_vertex in Gnx:
             compare_func = _compare_bfs_spc if return_sp_counter else \
-                           _compare_bfs
+                _compare_bfs
             compare_func(G, Gnx, start_vertex)
     else:  # Unknown type given to seed
         raise NotImplementedError("Invalid type for seed")
 
 
-def _compare_bfs(G,  Gnx, source):
+def _compare_bfs(G, Gnx, source):
     df = cugraph.bfs(G, source, return_sp_counter=False)
     # This call should only contain 3 columns:
     # 'vertex', 'distance', 'predecessor'
@@ -145,9 +125,9 @@ def _compare_bfs(G,  Gnx, source):
                            df['predecessor'].to_array())}
 
     nx_distances = nx.single_source_shortest_path_length(Gnx, source)
-    # TODO: The following only verifies vertices that were reached
+    # FIXME: The following only verifies vertices that were reached
     #       by cugraph's BFS.
-    # We assume that the distances are ginven back as integers in BFS
+    # We assume that the distances are given back as integers in BFS
     # max_val = np.iinfo(df['distance'].dtype).max
     # Unreached vertices have a distance of max_val